// frontend/src/app/services/models/tool-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ToolResponseDto {
  archived?: boolean;
  averageRating?: number;
  description?: string;
  id?: number;
  manufacturer?: string;
  name?: string;
  ownerName?: string;
  photo?: Array<string>;
  shareable?: boolean;
}